+++ examples/repeat.rs
@@ -27,6 +27,10 @@ fn setup(mut commands: Commands) {
                 if world.get::<ActionQueue>(agent).unwrap().is_empty() {
                     world.send_event(AppExit::Success);
                 }
+
+                // Do not advance action queue immediately,
+                // otherwise we get stuck in an infinite loop
+                // as we keep readding this action
                 false
             },
             repeat: u32::MAX,
@@ -52,18 +56,23 @@ impl<A: Action> Action for RepeatAction<A> {
         self.action.on_start(agent, world)
     }
 
-    fn on_stop(&mut self, agent: Entity, world: &mut World, reason: StopReason) {
+    fn on_stop(&mut self, agent: Option<Entity>, world: &mut World, reason: StopReason) {
         self.action.on_stop(agent, world, reason);
     }
 
-    fn on_drop(mut self: Box<Self>, agent: Entity, world: &mut World, reason: DropReason) {
+    fn on_remove(&mut self, agent: Option<Entity>, world: &mut World) {
+        self.action.on_remove(agent, world);
+    }
+
+    fn on_drop(mut self: Box<Self>, agent: Option<Entity>, world: &mut World, reason: DropReason) {
         if self.repeat == 0 || reason != DropReason::Done {
-            self.action.on_remove(agent, world);
             return;
         }
 
+        let Some(agent) = agent else { return };
+
         self.repeat -= 1;
-        world.get_mut::<ActionQueue>(agent).unwrap().push_back(self);
+        world.actions(agent).start(false).add(self as BoxedAction);
     }
 }
 
@@ -76,8 +85,8 @@ impl Action for PrintAction {
 
     fn on_start(&mut self, _agent: Entity, _world: &mut World) -> bool {
         println!("{}", self.0);
-        false
+        true
     }
 
-    fn on_stop(&mut self, _agent: Entity, _world: &mut World, _reason: StopReason) {}
+    fn on_stop(&mut self, _agent: Option<Entity>, _world: &mut World, _reason: StopReason) {}
 }