+++ src/traits.rs
@@ -1,10 +1,60 @@
-use downcast_rs::{impl_downcast, Downcast};
-
-use crate::*;
+use super::*;
 
 /// The trait that all actions must implement.
+///
+/// It contains various methods that together defines the _lifecycle_ of an action.
+/// From this, you can create any action that can last as long as you like,
+/// and do as much as you like.
+///
+/// In general, the act of _starting_ the action should happen inside [`on_start`](`Self::on_start`).
+/// This is especially true for despawning an `agent`.
+/// The other methods should be used for initializing and cleaning up.
+/// See below for more information.
+///
+/// ## ⚠️ Warning
+///
+/// Since you are given a mutable [`World`], you can in practice do _anything_.
+/// Depending on what you do, the logic for advancing the action queue might not work properly.
+///
+/// There are a few things you should keep in mind:
+///
+/// * If you want to despawn an `agent` as an action, this should be done in [`on_start`](`Self::on_start`).
+/// * The [`execute`](`ModifyActions::execute`) and [`next`](`ModifyActions::next`) methods should not be used,
+///     as that will immediately advance the action queue while inside any of the trait methods.
+///     Instead, you should return `true` in [`on_start`](`Self::on_start`).
+/// * When adding new actions, you should set the [`start`](`ModifyActions::start`) property to `false`.
+///     Otherwise, you will effectively call [`execute`](`ModifyActions::execute`) which, again, should not be used.
+///     At worst, you will cause a **stack overflow** if the action adds itself.
+///
+/// ```rust,no_run
+/// # use bevy_ecs::prelude::*;
+/// # use bevy_sequential_actions::*;
+/// # struct EmptyAction;
+/// # impl Action for EmptyAction {
+/// #   fn is_finished(&self, _a: Entity, _w: &World) -> bool { true }
+/// #   fn on_start(&mut self, _a: Entity, _w: &mut World) -> bool { true }
+/// #   fn on_stop(&mut self, _a: Option<Entity>, _w: &mut World, _r: StopReason) {}
+/// # }
+/// # struct TestAction;
+/// # impl Action for TestAction {
+/// #   fn is_finished(&self, _a: Entity, _w: &World) -> bool { true }
+///     fn on_start(&mut self, agent: Entity, world: &mut World) -> bool {
+/// #       let action_a = EmptyAction;
+/// #       let action_b = EmptyAction;
+/// #       let action_c = EmptyAction;
+///         world
+///             .actions(agent)
+///             .start(false) // Do not start next action
+///             .add_many(actions![action_a, action_b, action_c]);
+///
+///         // Immediately advance the action queue
+///         true
+///     }
+/// #   fn on_stop(&mut self, _a: Option<Entity>, _w: &mut World, _r: StopReason) {}
+/// # }
+/// ```
 #[allow(unused_variables)]
-pub trait Action: Downcast + Send + Sync + 'static {
+pub trait Action: downcast_rs::Downcast + Send + Sync + 'static {
     /// Determines if an action is finished or not.
     /// Advances the action queue when returning `true`.
     ///
@@ -13,22 +63,38 @@ pub trait Action: Downcast + Send + Sync + 'static {
 
     /// The method that is called when an action is started.
     ///
+    /// Typically here you would insert components to `agent` or a new entity
+    /// to make systems run. If you wish to despawn `agent` as an action,
+    /// **this is where you should do it**.
+    ///
     /// Returning `true` here marks the action as already finished,
     /// and will immediately advance the action queue.
     fn on_start(&mut self, agent: Entity, world: &mut World) -> bool;
 
     /// The method that is called when an action is stopped.
-    fn on_stop(&mut self, agent: Entity, world: &mut World, reason: StopReason);
+    ///
+    /// Typically here you would clean up any stuff from [`on_start`](`Self::on_start`),
+    /// depending on the [`reason`](`StopReason`).
+    ///
+    /// When paused, the action will be put back into the action queue again to the front.
+    /// This means that it will start again when the next action is executed.
+    fn on_stop(&mut self, agent: Option<Entity>, world: &mut World, reason: StopReason);
 
     /// The method that is called when an action is added to the queue.
+    ///
+    /// You can think of this as the _constructor_, as it will only be called once.
     fn on_add(&mut self, agent: Entity, world: &mut World) {}
 
     /// The method that is called when an action is removed from the queue.
-    fn on_remove(&mut self, agent: Entity, world: &mut World) {}
+    ///
+    /// You can think of this as the _destructor_, as it will only be called once.
+    fn on_remove(&mut self, agent: Option<Entity>, world: &mut World) {}
 
-    /// The last method that is called for an action.
-    /// Full ownership is given here, hence the name.
-    fn on_drop(self: Box<Self>, agent: Entity, world: &mut World, reason: DropReason) {}
+    /// The last method that is called with full ownership.
+    ///
+    /// This is useful for actions that might want to alter the behavior of the action queue.
+    /// For example, a `RepeatAction` could keep readding itself to the action queue based on some counter.
+    fn on_drop(self: Box<Self>, agent: Option<Entity>, world: &mut World, reason: DropReason) {}
 
     /// Returns the type name of an action.
     fn type_name(&self) -> &'static str {
@@ -36,7 +102,7 @@ pub trait Action: Downcast + Send + Sync + 'static {
     }
 }
 
-impl_downcast!(Action);
+downcast_rs::impl_downcast!(Action);
 
 impl<T> From<T> for BoxedAction
 where
@@ -59,7 +125,7 @@ where
         (self)(agent, world)
     }
 
-    fn on_stop(&mut self, _agent: Entity, _world: &mut World, _reason: StopReason) {}
+    fn on_stop(&mut self, _agent: Option<Entity>, _world: &mut World, _reason: StopReason) {}
 }
 
 impl std::fmt::Debug for BoxedAction {
@@ -93,10 +159,11 @@ pub trait ModifyActions {
     fn add(&mut self, action: impl Into<BoxedAction>) -> &mut Self;
 
     /// Adds a collection of actions to the queue.
+    /// An empty collection does nothing.
     fn add_many<I>(&mut self, actions: I) -> &mut Self
     where
         I: IntoIterator<Item = BoxedAction> + Send + 'static,
-        I::IntoIter: DoubleEndedIterator;
+        I::IntoIter: DoubleEndedIterator + ExactSizeIterator + Debug;
 
     /// [`Starts`](Action::on_start) the next [`action`](Action) in the queue,
     /// but only if there is no current action.